--- tests/Rector/ReturnArrayClassMethodToYieldRector/Source/EventSubscriberInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EonX\EasyQuality\Tests\Rector\ReturnArrayClassMethodToYieldRector\Source;
-
-interface EventSubscriberInterface
-{
-
-}
