+++ tests/Rector/ReturnArrayToYieldRector/Source/EventSubscriberInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyQuality\Tests\Rector\ReturnArrayToYieldRector\Source;
+
+interface EventSubscriberInterface
+{
+
+}
